requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0
scikit-learn>=1.3.0
selenium>=4.0.0
webdriver-manager>=4.0.0
